Include all platform configs in platform object
Turn all the platform specific sources in to headers with data only.
